requirements.txt
braintree==3.31.0
Django==1.10.3
django-braces==1.10.0
django-cors-middleware==1.3.1
django-oauth-toolkit==0.10.0
djangorestframework==3.5.3
httpie==0.9.6
oauthlib==1.0.3
Pygments==2.1.3
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
